Move More Types from store.rs to types.rs.
I think there are actually three distinct groups of types here:

1. types.rs (domain types, shared).
2. options.rs (input types)
3. results.rs (output types)

Will start on that in the next commit.
